backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-socketio==5.11.1
groq==0.9.0
elevenlabs==1.9.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.2
websockets==13.1